requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
filetype==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Js2Py_3.13==0.74.1
lxml==6.0.0
m3u8==6.0.0
numpy==2.3.1
pandas==2.3.1
pyjsparser==2.7.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
twikit==2.3.3
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
webvtt-py==0.5.1
